Add implementation for 4 way dmux.
It is hard to explain how I get this... just that we use a first 2 way
dmux to create two signals and then use another select line to create
another 2 way dmux.

So one dmux outputs a and c while another outputs b and d. Its not
intutive to me why that must be the case but this logic can easily be
figured out on doing.